Adds the styling using vanilla-extract and a basic scene
This adds a the reset of the styling to use in the project along with some base styles needed. It also creates a basic scene to show and test. Vanilla extract was extended with some missing properties and eslint change to turn off an issue with unknown properties, is a known issue in three and can't be fixed as my result, so I just turn it off.