Improves the scene, the styles, groups routes for the website and add menu state
Improves a couple things here and there for the scene, still just for testing. Improves styling for the scene and the navigation bar. It groups all the website routes in a group to share the same layout while the game it's left alone to a avoid any game menu showing up in the website. And finally ignores the lock file in the format and liniting rules. It also changes the tsconfig target since vanilla extract requires modern js