Extracts the game menu to its own module
Since the game menu will be huge in the future it was extracted to its own module. Testing naming was also change in all of the test suits. The dependency '@testing-library/react-hooks' was also not being used and when I tried to used was outdated so I decided I was better and simple just to use '@testing-library/react' to test hooks as well.