Removes warning from three.js while testing
Removes the warning and adds also a way to remove other warnings in the future. Right now Three.js has a warning for importing multiple times the instance. Testing does not affect at all and trying to remove the issue is a bigger problem for something that it won't be causing too much issues. Therefore I decided to turn off the warning in the log scope. By just removing the log if is the warning.